+++ src/platform_impl/windows/status_bar.rs
@@ -3,7 +3,7 @@ use crate::{
   error::OsError,
   menu::{MenuItem, MenuType},
   platform_impl::EventLoopWindowTarget,
-  status_bar::Statusbar as RootStatusbar,
+  platform::windows::Statusbar as RootStatusbar,
 };
 use std::cell::RefCell;
 use winapi::{
